+++ scripts/rollback/rollback-2.13.0.sh
@@ -1,0 +1,186 @@
+#!/bin/bash
+# timestamp: 1630997684975
+
+npm dist-tag add @midwayjs/egg-layer@2.12.9 latest
+npm dist-tag add @midwayjs/egg-layer@2.13.0 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.12.9 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.13.0 latest
+npm dist-tag add @midwayjs/express-layer@2.12.9 latest
+npm dist-tag add @midwayjs/express-layer@2.13.0 beta
+tnpm dist-tag add @midwayjs/express-layer@2.12.9 latest
+tnpm dist-tag add @midwayjs/express-layer@2.13.0 latest
+npm dist-tag add @midwayjs/koa-layer@2.12.9 latest
+npm dist-tag add @midwayjs/koa-layer@2.13.0 beta
+tnpm dist-tag add @midwayjs/koa-layer@2.12.9 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.13.0 latest
+npm dist-tag add @midwayjs/serverless-app@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-app@2.13.0 beta
+tnpm dist-tag add @midwayjs/serverless-app@2.12.9 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.13.0 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.13.0 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.12.9 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.13.0 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.13.0 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.12.9 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.13.0 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.13.0 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.12.9 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.13.0 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@2.13.0 beta
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@2.12.9 latest
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@2.13.0 latest
+npm dist-tag add @midwayjs/static-layer@2.12.9 latest
+npm dist-tag add @midwayjs/static-layer@2.13.0 beta
+tnpm dist-tag add @midwayjs/static-layer@2.12.9 latest
+tnpm dist-tag add @midwayjs/static-layer@2.13.0 latest
+npm dist-tag add @midwayjs/axios@2.12.9 latest
+npm dist-tag add @midwayjs/axios@2.13.0 beta
+tnpm dist-tag add @midwayjs/axios@2.12.9 latest
+tnpm dist-tag add @midwayjs/axios@2.13.0 latest
+npm dist-tag add @midwayjs/bootstrap@2.12.9 latest
+npm dist-tag add @midwayjs/bootstrap@2.13.0 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.12.9 latest
+tnpm dist-tag add @midwayjs/bootstrap@2.13.0 latest
+npm dist-tag add @midwayjs/cache@2.12.9 latest
+npm dist-tag add @midwayjs/cache@2.13.0 beta
+tnpm dist-tag add @midwayjs/cache@2.12.9 latest
+tnpm dist-tag add @midwayjs/cache@2.13.0 latest
+npm dist-tag add @midwayjs/consul@2.12.9 latest
+npm dist-tag add @midwayjs/consul@2.13.0 beta
+tnpm dist-tag add @midwayjs/consul@2.12.9 latest
+tnpm dist-tag add @midwayjs/consul@2.13.0 latest
+npm dist-tag add @midwayjs/core@2.12.9 latest
+npm dist-tag add @midwayjs/core@2.13.0 beta
+tnpm dist-tag add @midwayjs/core@2.12.9 latest
+tnpm dist-tag add @midwayjs/core@2.13.0 latest
+npm dist-tag add @midwayjs/cos@2.12.9 latest
+npm dist-tag add @midwayjs/cos@2.13.0 beta
+tnpm dist-tag add @midwayjs/cos@2.12.9 latest
+tnpm dist-tag add @midwayjs/cos@2.13.0 latest
+npm dist-tag add @midwayjs/decorator@2.12.9 latest
+npm dist-tag add @midwayjs/decorator@2.13.0 beta
+tnpm dist-tag add @midwayjs/decorator@2.12.9 latest
+tnpm dist-tag add @midwayjs/decorator@2.13.0 latest
+npm dist-tag add @midwayjs/faas@2.12.9 latest
+npm dist-tag add @midwayjs/faas@2.13.0 beta
+tnpm dist-tag add @midwayjs/faas@2.12.9 latest
+tnpm dist-tag add @midwayjs/faas@2.13.0 latest
+npm dist-tag add @midwayjs/grpc@2.12.9 latest
+npm dist-tag add @midwayjs/grpc@2.13.0 beta
+tnpm dist-tag add @midwayjs/grpc@2.12.9 latest
+tnpm dist-tag add @midwayjs/grpc@2.13.0 latest
+npm dist-tag add @midwayjs/logger@2.11.3 latest
+npm dist-tag add @midwayjs/logger@2.13.0 beta
+tnpm dist-tag add @midwayjs/logger@2.11.3 latest
+tnpm dist-tag add @midwayjs/logger@2.13.0 latest
+npm dist-tag add midway-schedule@2.12.9 latest
+npm dist-tag add midway-schedule@2.13.0 beta
+tnpm dist-tag add midway-schedule@2.12.9 latest
+tnpm dist-tag add midway-schedule@2.13.0 latest
+npm dist-tag add midway@2.12.9 latest
+npm dist-tag add midway@2.13.0 beta
+tnpm dist-tag add midway@2.12.9 latest
+tnpm dist-tag add midway@2.13.0 latest
+npm dist-tag add @midwayjs/mock@2.12.9 latest
+npm dist-tag add @midwayjs/mock@2.13.0 beta
+tnpm dist-tag add @midwayjs/mock@2.12.9 latest
+tnpm dist-tag add @midwayjs/mock@2.13.0 latest
+npm dist-tag add @midwayjs/orm@2.12.9 latest
+npm dist-tag add @midwayjs/orm@2.13.0 beta
+tnpm dist-tag add @midwayjs/orm@2.12.9 latest
+tnpm dist-tag add @midwayjs/orm@2.13.0 latest
+npm dist-tag add @midwayjs/oss@2.12.9 latest
+npm dist-tag add @midwayjs/oss@2.13.0 beta
+tnpm dist-tag add @midwayjs/oss@2.12.9 latest
+tnpm dist-tag add @midwayjs/oss@2.13.0 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@2.12.9 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@2.13.0 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.12.9 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.13.0 latest
+npm dist-tag add @midwayjs/prometheus@2.12.9 latest
+npm dist-tag add @midwayjs/prometheus@2.13.0 beta
+tnpm dist-tag add @midwayjs/prometheus@2.12.9 latest
+tnpm dist-tag add @midwayjs/prometheus@2.13.0 latest
+npm dist-tag add @midwayjs/rabbitmq@2.12.9 latest
+npm dist-tag add @midwayjs/rabbitmq@2.13.0 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.12.9 latest
+tnpm dist-tag add @midwayjs/rabbitmq@2.13.0 latest
+npm dist-tag add @midwayjs/redis@2.12.9 latest
+npm dist-tag add @midwayjs/redis@2.13.0 beta
+tnpm dist-tag add @midwayjs/redis@2.12.9 latest
+tnpm dist-tag add @midwayjs/redis@2.13.0 latest
+npm dist-tag add @midwayjs/socketio@2.12.9 latest
+npm dist-tag add @midwayjs/socketio@2.13.0 beta
+tnpm dist-tag add @midwayjs/socketio@2.12.9 latest
+tnpm dist-tag add @midwayjs/socketio@2.13.0 latest
+npm dist-tag add @midwayjs/task@2.12.9 latest
+npm dist-tag add @midwayjs/task@2.13.0 beta
+tnpm dist-tag add @midwayjs/task@2.12.9 latest
+tnpm dist-tag add @midwayjs/task@2.13.0 latest
+npm dist-tag add @midwayjs/typegoose@2.12.9 latest
+npm dist-tag add @midwayjs/typegoose@2.13.0 beta
+tnpm dist-tag add @midwayjs/typegoose@2.12.9 latest
+tnpm dist-tag add @midwayjs/typegoose@2.13.0 latest
+npm dist-tag add @midwayjs/version@2.12.9 latest
+npm dist-tag add @midwayjs/version@2.13.0 beta
+tnpm dist-tag add @midwayjs/version@2.12.9 latest
+tnpm dist-tag add @midwayjs/version@2.13.0 latest
+npm dist-tag add @midwayjs/express@2.12.9 latest
+npm dist-tag add @midwayjs/express@2.13.0 beta
+tnpm dist-tag add @midwayjs/express@2.12.9 latest
+tnpm dist-tag add @midwayjs/express@2.13.0 latest
+npm dist-tag add @midwayjs/koa@2.12.9 latest
+npm dist-tag add @midwayjs/koa@2.13.0 beta
+tnpm dist-tag add @midwayjs/koa@2.12.9 latest
+tnpm dist-tag add @midwayjs/koa@2.13.0 latest
+npm dist-tag add @midwayjs/web@2.12.9 latest
+npm dist-tag add @midwayjs/web@2.13.0 beta
+tnpm dist-tag add @midwayjs/web@2.12.9 latest
+tnpm dist-tag add @midwayjs/web@2.13.0 latest
+npm dist-tag add @midwayjs/ws@2.12.9 latest
+npm dist-tag add @midwayjs/ws@2.13.0 beta
+tnpm dist-tag add @midwayjs/ws@2.12.9 latest
+tnpm dist-tag add @midwayjs/ws@2.13.0 latest
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.12.9 => 2.13.0
+#  - @midwayjs/express-layer: 2.12.9 => 2.13.0
+#  - @midwayjs/koa-layer: 2.12.9 => 2.13.0
+#  - @midwayjs/serverless-app: 2.12.9 => 2.13.0
+#  - @midwayjs/serverless-fc-starter: 2.12.9 => 2.13.0
+#  - @midwayjs/serverless-fc-trigger: 2.12.9 => 2.13.0
+#  - @midwayjs/serverless-scf-trigger: 2.12.9 => 2.13.0
+#  - @midwayjs/serverless-vercel-starter: 2.12.9 => 2.13.0
+#  - @midwayjs/static-layer: 2.12.9 => 2.13.0
+#  - @midwayjs/axios: 2.12.9 => 2.13.0
+#  - @midwayjs/bootstrap: 2.12.9 => 2.13.0
+#  - @midwayjs/cache: 2.12.9 => 2.13.0
+#  - @midwayjs/consul: 2.12.9 => 2.13.0
+#  - @midwayjs/core: 2.12.9 => 2.13.0
+#  - @midwayjs/cos: 2.12.9 => 2.13.0
+#  - @midwayjs/decorator: 2.12.9 => 2.13.0
+#  - @midwayjs/faas: 2.12.9 => 2.13.0
+#  - @midwayjs/grpc: 2.12.9 => 2.13.0
+#  - @midwayjs/logger: 2.11.3 => 2.13.0
+#  - midway-schedule: 2.12.9 => 2.13.0
+#  - midway: 2.12.9 => 2.13.0
+#  - @midwayjs/mock: 2.12.9 => 2.13.0
+#  - @midwayjs/orm: 2.12.9 => 2.13.0
+#  - @midwayjs/oss: 2.12.9 => 2.13.0
+#  - @midwayjs/prometheus-socket-io: 2.12.9 => 2.13.0
+#  - @midwayjs/prometheus: 2.12.9 => 2.13.0
+#  - @midwayjs/rabbitmq: 2.12.9 => 2.13.0
+#  - @midwayjs/redis: 2.12.9 => 2.13.0
+#  - @midwayjs/socketio: 2.12.9 => 2.13.0
+#  - @midwayjs/task: 2.12.9 => 2.13.0
+#  - @midwayjs/typegoose: 2.12.9 => 2.13.0
+#  - @midwayjs/version: 2.12.9 => 2.13.0
+#  - @midwayjs/express: 2.12.9 => 2.13.0
+#  - @midwayjs/koa: 2.12.9 => 2.13.0
+#  - @midwayjs/web: 2.12.9 => 2.13.0
+#  - @midwayjs/ws: 2.12.9 => 2.13.0