+++ packages-serverless/koa-layer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/midwayjs/midway-faas/compare/v2.12.9...v2.13.0) (2021-09-07)
+
+**Note:** Version bump only for package @midwayjs/koa-layer
+
+
+
+
+
 ## [2.12.9](https://github.com/midwayjs/midway-faas/compare/v2.12.8...v2.12.9) (2021-09-01)
 
 **Note:** Version bump only for package @midwayjs/koa-layer