+++ packages/axios/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/midwayjs/midway/compare/v2.12.9...v2.13.0) (2021-09-07)
+
+
+### Features
+
+* add redis component ([#1270](https://github.com/midwayjs/midway/issues/1270)) ([cea81a8](https://github.com/midwayjs/midway/commit/cea81a8a1a775be90863b40616f8b86d0df6ec06))
+
+
+
+
+
 ## [2.12.9](https://github.com/midwayjs/midway/compare/v2.12.8...v2.12.9) (2021-09-01)
 
 **Note:** Version bump only for package @midwayjs/axios