+++ packages/logger/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/midwayjs/midway/compare/v2.12.9...v2.13.0) (2021-09-07)
+
+
+### Bug Fixes
+
+* node v16 typings ([#1269](https://github.com/midwayjs/midway/issues/1269)) ([59d0ee7](https://github.com/midwayjs/midway/commit/59d0ee71215587fd22b3627faa5c1bf26b1dd0e2))
+* winston rotate log size limit in cluster mode ([#1268](https://github.com/midwayjs/midway/issues/1268)) ([3299506](https://github.com/midwayjs/midway/commit/32995064f05955ab4d438741d056e819481ef30e))
+
+
+
+
+
 ## [2.11.3](https://github.com/midwayjs/midway/compare/v2.11.2...v2.11.3) (2021-07-02)